refactor: api controller와 페이지 반환 controller 분리 및 `@RestController`와 `@RequestMapping` 사용
