+++ src/main/java/roomescape/presentation/api/ReservationController.java
@@ -1,14 +1,14 @@
-package roomescape.presentation;
+package roomescape.presentation.api;
 
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.application.ReservationService;
 import roomescape.application.dto.CreateInfoReservationDto;
 import roomescape.application.dto.CreateReservationDto;
@@ -20,7 +20,8 @@ import roomescape.presentation.dto.response.ReadReservationResponse;
 import java.net.URI;
 import java.util.List;
 
-@Controller
+@RestController
+@RequestMapping("/reservations")
 public class ReservationController {
 
     private final ReservationService reservationService;
@@ -29,13 +30,7 @@ public class ReservationController {
         this.reservationService = reservationService;
     }
 
-    @GetMapping("/reservation")
-    public String reservation() {
-        return "new-reservation.html";
-    }
-
-    @GetMapping("/reservations")
-    @ResponseBody
+    @GetMapping
     public ResponseEntity<List<ReadReservationResponse>> readAll() {
         final List<ReadReservationDto> readReservationDtos = reservationService.readAll();
         final List<ReadReservationResponse> responses = readReservationDtos.stream()
@@ -45,8 +40,7 @@ public class ReservationController {
         return ResponseEntity.ok(responses);
     }
 
-    @PostMapping("/reservations")
-    @ResponseBody
+    @PostMapping
     public ResponseEntity<CreateReservationResponse> create(@RequestBody @Valid final CreateReservationRequest request) {
         final CreateReservationDto createReservationDto = new CreateReservationDto(
                 request.getName(),
@@ -60,7 +54,7 @@ public class ReservationController {
                              .body(response);
     }
 
-    @DeleteMapping("/reservations/{id}")
+    @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable final Long id) {
         reservationService.deleteById(id);
