+++ src/test/java/roomescape/MissionStepTest.java
@@ -13,7 +13,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
 import roomescape.domain.Reservation;
-import roomescape.presentation.ReservationController;
+import roomescape.presentation.api.ReservationController;
 
 import java.lang.reflect.Field;
 import java.sql.Connection;
